WIP: Convert tests to TypeScript and update them for new API
This is probably doing a bit too much at once, but so be it. We changed 
the API and broke the tests, but I found it annoying to not have types 
in tests since it would be nice to have the test exercise the types in 
the API in addition to the behavior. So I converted to TypeScript.

After trying multiple solutions, the only thing I could get to work is 
processing the tests file with webpack. We need webpack anyway to 
process the output of wasm_bindgen and compile the TypeScript in the 
implementation, so it's not a big addition to just compile the tests 
with it as well. I added an environment variable to the webpack config 
so we can just process the tests entrypoint when running `yarn test`.